// src/components/utils/enums/filter.ts
export enum DataType {
  STRING = 'string',
  FLOAT = 'float',
  INT = 'int',
  DATE = 'date',
  BOOL = 'boolean',
  DATETIME = 'datetime',
  JSON = 'json',
}

export enum FilterOperator {
  IS = 'is',
  LIKE = 'like',
}

export enum FilterOtpion {
  DROPDOWN = 'dropdown',
  INPUT = 'input',
}